requirements.txt
asgiref==3.7.2
Django==4.2
django-cors-headers==4.2.0
django-environ==0.10.0
djangorestframework==3.14.0
drf-yasg==1.21.7
flake8==6.0.0
google-api-python-client==2.97.0
inflection==0.5.1
mccabe==0.7.0
mypy==1.4.1
mypy-extensions==1.0.0
packaging==23.1
psycopg2-binary~=2.9.3
pycodestyle==2.10.0
pyflakes==3.0.1
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
uritemplate==4.1.1
